--- components/Icons/FirebaseLogo.tsx
@@ -1,16 +1,0 @@
-import React from "react";
-import { Icon, IconProps } from "@chakra-ui/react";
-
-function FirebaseIcon(props: IconProps): JSX.Element {
-  return (
-    <Icon viewBox="0 0 50 68" {...props}>
-      <path
-        fill="currentColor"
-        d="M1.6066989,43.8732883 L8.22720994,1.29197398 C8.32777924,0.643112236 8.82990599,0.131546879 9.47412157,0.0216238467 C10.1183371,-0.0882991861 10.7603352,0.228043208 11.0683702,0.807184472 L18.1864641,14.1675251 L1.6066989,43.8732883 Z M48.6187845,54.2065575 L42.3173343,15.0187697 C42.2221419,14.4559508 41.8224842,13.9935122 41.2812968,13.8199848 C40.7401094,13.6464574 40.147804,13.7908309 39.7458564,14.1942458 L0,54.2065575 L21.9924033,66.6011364 C23.3734166,67.3760485 25.0554508,67.3760485 26.4364641,66.6011364 L48.6187845,54.2065575 Z M30.7475829,20.0079026 L25.6540055,10.2624884 C25.3913218,9.76069235 24.8735865,9.44650145 24.3093923,9.44650145 C23.7451981,9.44650145 23.2274628,9.76069235 22.964779,10.2624884 L0.600138122,50.3435104 L30.7475829,20.0079026 Z"
-        id="icon"
-      />
-    </Icon>
-  );
-}
-
-export default FirebaseIcon;
