--- components/Icons/PythonIcon.tsx
@@ -1,19 +1,0 @@
-import React from "react";
-import { Icon, IconProps } from "@chakra-ui/react";
-
-function PythonIcon(props: IconProps): JSX.Element {
-  return (
-    <Icon viewBox="0 0 51 51" {...props}>
-      <path
-        d="M31.2774 3.62322C29.4256 3.31841 26.7152 3.17796 24.8755 3.18693C23.151 3.18892 21.4298 3.33486 19.7296 3.62322C15.1735 4.42009 14.3438 6.08357 14.3438 9.15552V12.7494H25.5V14.3432H10.72C6.58021 14.3432 3.22039 19.2908 3.18751 25.4118C3.18751 25.4406 3.18751 25.4695 3.18751 25.4994C3.18608 26.5696 3.29051 27.6373 3.49929 28.6869C4.42466 33.2968 7.30736 36.6557 10.72 36.6557H12.75V31.2768C12.75 28.159 14.8259 25.3052 18.2285 24.6926L19.2117 24.5701H31.2774C31.7841 24.572 32.2888 24.505 32.7775 24.3708C33.8947 24.0725 34.8815 23.4121 35.5834 22.4931C36.2853 21.574 36.6626 20.4482 36.6563 19.2918V9.15552C36.6563 6.27482 34.1979 4.10234 31.2774 3.62322ZM19.4169 10.5082C19.015 10.5096 18.6217 10.3916 18.287 10.1692C17.9522 9.94674 17.6911 9.62992 17.5366 9.25888C17.3822 8.88784 17.3414 8.47929 17.4195 8.08504C17.4976 7.6908 17.691 7.32862 17.9752 7.04443C18.2593 6.76025 18.6215 6.56686 19.0158 6.48879C19.41 6.41072 19.8186 6.45149 20.1896 6.60593C20.5606 6.76036 20.8775 7.02151 21.0999 7.35625C21.3223 7.69099 21.4403 8.08425 21.4389 8.48614C21.4391 8.75172 21.3869 9.01472 21.2853 9.26011C21.1837 9.5055 21.0348 9.72846 20.847 9.91625C20.6592 10.104 20.4362 10.253 20.1908 10.3546C19.9454 10.4561 19.6824 10.5083 19.4169 10.5082Z"
-        fill="currentColor"
-      />
-      <path
-        d="M47.3423 21.6172C46.2765 17.3728 43.5163 14.3457 40.28 14.3457H38.516V19.0722C38.516 23.0137 35.9261 25.8675 32.7645 26.3566C32.4451 26.4061 32.1225 26.4311 31.7993 26.4313H19.7306C19.2187 26.4308 18.7089 26.4961 18.2136 26.6255C15.9913 27.2093 14.3438 29.058 14.3438 31.4556V41.5928C14.3438 44.4815 17.2464 46.1749 20.0942 47.0026C23.5018 47.9937 27.2143 48.173 31.2923 47.0026C34.0007 46.2287 36.6562 44.6638 36.6562 41.5928V38.2519H25.5V36.6582H40.28C43.2125 36.6582 45.7536 34.1749 46.9987 30.5511C47.5455 28.9242 47.8204 27.2183 47.8125 25.502C47.8153 24.1927 47.6574 22.888 47.3423 21.6172ZM31.5274 40.2441C31.9293 40.2428 32.3225 40.3608 32.6573 40.5832C32.992 40.8056 33.2531 41.1224 33.4076 41.4935C33.562 41.8645 33.6028 42.2731 33.5247 42.6673C33.4466 43.0615 33.2533 43.4237 32.9691 43.7079C32.6849 43.9921 32.3227 44.1855 31.9285 44.2636C31.5342 44.3416 31.1257 44.3009 30.7546 44.1464C30.3836 43.992 30.0668 43.7308 29.8443 43.3961C29.6219 43.0614 29.5039 42.6681 29.5053 42.2662C29.5052 42.0006 29.5574 41.7376 29.6589 41.4922C29.7605 41.2468 29.9095 41.0239 30.0972 40.8361C30.285 40.6483 30.508 40.4994 30.7534 40.3978C30.9988 40.2962 31.2618 40.244 31.5274 40.2441Z"
-        fill="currentColor"
-      />
-    </Icon>
-  );
-}
-
-export default PythonIcon;
