--- components/Icons/ReactIcon.tsx
@@ -1,19 +1,0 @@
-import React from "react";
-import { Icon, IconProps } from "@chakra-ui/react";
-
-function ReactIcon(props: IconProps): JSX.Element {
-  return (
-    <Icon viewBox="0 0 52 52" {...props}>
-      <path
-        d="M41.7086 18.3574C41.1893 18.18 40.6663 18.0148 40.1395 17.8618C40.2268 17.5063 40.3074 17.1488 40.3812 16.7893C41.5695 11.0104 40.7925 6.35477 38.1407 4.82219C35.6017 3.35156 31.4376 4.88414 27.237 8.54851C26.824 8.91008 26.4194 9.28247 26.0233 9.6657C25.7593 9.41112 25.4884 9.16195 25.2108 8.9182C20.8081 4.99992 16.3952 3.34953 13.7464 4.88719C11.2074 6.36086 10.4528 10.7372 11.5222 16.2145C11.6292 16.7561 11.7494 17.2954 11.8828 17.8323C11.2582 18.0101 10.6549 18.2 10.078 18.4011C4.92269 20.2038 1.62598 23.0303 1.62598 25.9614C1.62598 28.99 5.16441 32.0267 10.5401 33.868C10.9762 34.0163 11.4163 34.1541 11.8604 34.2814C11.7162 34.8563 11.5886 35.4375 11.4775 36.0252C10.4619 41.408 11.2541 45.6808 13.7881 47.1453C16.4053 48.6576 20.7959 47.1037 25.0757 43.356C25.4143 43.0595 25.7528 42.7463 26.0914 42.4166C26.5193 42.8303 26.9594 43.2318 27.4117 43.6211C31.5554 47.1941 35.6474 48.6373 38.1773 47.1687C40.7925 45.6513 41.6426 41.0607 40.5386 35.4758C40.454 35.0492 40.3565 34.6139 40.2461 34.1697C40.5549 34.0783 40.8572 33.9838 41.1531 33.8863C46.739 32.0308 50.378 29.0306 50.378 25.9645C50.376 23.0191 46.9736 20.1744 41.7086 18.3574ZM28.6071 10.1227C32.2054 6.98344 35.5692 5.74437 37.1028 6.63C38.7349 7.57352 39.3696 11.378 38.3439 16.3678C38.2761 16.6928 38.203 17.0168 38.1245 17.3398C35.9771 16.8513 33.7995 16.5066 31.6062 16.3079C30.3504 14.4989 28.9738 12.7769 27.4858 11.1536C27.8501 10.8002 28.2232 10.4565 28.605 10.1227H28.6071ZM16.0204 28.4659C16.4612 29.3191 16.922 30.1617 17.4027 30.9938C17.8922 31.8429 18.4024 32.6794 18.9332 33.5034C17.4241 33.3394 15.9244 33.0977 14.4401 32.7793C14.8707 31.3859 15.4009 29.9386 16.0204 28.4659ZM16.0204 23.5574C15.4111 22.1162 14.8951 20.6984 14.4716 19.3294C15.863 19.0176 17.3458 18.7627 18.8916 18.5697C18.3736 19.3781 17.8753 20.1984 17.3966 21.0305C16.9179 21.8627 16.4582 22.705 16.0174 23.5574H16.0204ZM17.1275 26.0122C17.7693 24.6736 18.4606 23.3607 19.2014 22.0736C19.9414 20.7871 20.7292 19.5301 21.5647 18.3026C23.015 18.1929 24.5019 18.135 26.004 18.135C27.5061 18.135 29.0021 18.1929 30.4514 18.3036C31.2782 19.5284 32.0612 20.781 32.8006 22.0614C33.54 23.3418 34.2404 24.6485 34.9019 25.9817C34.2492 27.3251 33.5504 28.6433 32.8057 29.9366C32.067 31.223 31.2883 32.4838 30.4697 33.7188C29.0225 33.8203 27.5244 33.8762 26.001 33.8762C24.4775 33.8762 23.0079 33.8284 21.5871 33.736C20.7455 32.5044 19.9512 31.242 19.2044 29.9487C18.4576 28.6555 17.7643 27.3433 17.1244 26.0122H17.1275ZM34.6084 30.9796C35.0993 30.1278 35.5705 29.2649 36.0221 28.3908C36.6412 29.7945 37.1877 31.2291 37.6593 32.6889C36.1576 33.0271 34.6392 33.2864 33.1103 33.4659C33.6263 32.6473 34.1256 31.8185 34.6084 30.9796ZM36.0018 23.5584C35.5522 22.7019 35.0847 21.8556 34.5993 21.0194C34.1253 20.1947 33.6314 19.3802 33.1175 18.5758C34.6724 18.7728 36.1643 19.0348 37.5628 19.3548C37.1132 20.7812 36.5922 22.1842 36.0018 23.5584ZM26.0243 12.6455C27.0385 13.7537 27.9961 14.9123 28.8935 16.117C26.9739 16.0255 25.0531 16.0255 23.1308 16.117C24.0784 14.8637 25.0483 13.7008 26.0243 12.6455ZM14.7946 6.69195C16.4257 5.74539 20.0321 7.0982 23.8336 10.4772C24.0764 10.6935 24.3211 10.92 24.5649 11.1546C23.0692 12.7777 21.6829 14.4982 20.415 16.3048C18.2266 16.502 16.0533 16.8413 13.9089 17.3205C13.785 16.8221 13.6733 16.3204 13.5738 15.8153C12.6546 11.1201 13.263 7.58062 14.7946 6.69195ZM12.417 32.2705C12.0107 32.1547 11.6096 32.0291 11.2135 31.8937C8.83691 31.0812 6.87371 30.0209 5.52598 28.8661C4.31941 27.8312 3.70801 26.7983 3.70801 25.9614C3.70801 24.183 6.35473 21.9141 10.7686 20.3755C11.3232 20.1825 11.8824 20.0061 12.4464 19.8463C13.102 21.953 13.8959 24.0142 14.823 26.0162C13.8848 28.046 13.0811 30.1353 12.417 32.2705ZM23.7036 41.7838C21.8115 43.4413 19.9164 44.6164 18.2457 45.2085C16.7446 45.7397 15.5492 45.7549 14.8271 45.3375C13.2894 44.4488 12.6496 41.017 13.522 36.4142C13.6256 35.8726 13.7441 35.3309 13.8775 34.7892C16.0433 35.2558 18.2384 35.5744 20.4475 35.7429C21.7278 37.5589 23.1256 39.2892 24.6319 40.9226C24.3293 41.2202 24.0195 41.5066 23.7036 41.7838ZM26.0822 39.4255C25.095 38.3581 24.1099 37.177 23.1491 35.9074C24.0828 35.944 25.0334 35.9623 26.001 35.9623C26.9943 35.9623 27.9774 35.9409 28.9463 35.8973C28.0521 37.1212 27.0962 38.2989 26.0822 39.4255ZM38.7288 42.3292C38.4383 43.8973 37.8543 44.9434 37.1322 45.3619C35.5956 46.2536 32.31 45.0948 28.7675 42.0388C28.3613 41.6894 27.955 41.3146 27.5417 40.9216C29.0187 39.2818 30.3837 37.5446 31.6275 35.7216C33.8489 35.5334 36.0546 35.1923 38.2291 34.7009C38.3286 35.103 38.4166 35.4971 38.4932 35.883C38.9817 38.351 39.0518 40.5834 38.7288 42.3292ZM40.496 31.9048C40.2299 31.9932 39.9557 32.0785 39.6774 32.1618C38.9955 30.0452 38.1706 27.9773 37.2084 25.9726C38.1345 23.9946 38.9272 21.9567 39.5809 19.8727C40.0796 20.018 40.564 20.1703 41.0302 20.3318C45.5416 21.8877 48.2939 24.1912 48.2939 25.9614C48.2939 27.8505 45.3212 30.3022 40.496 31.9048Z"
-        fill="currentColor"
-      />
-      <path
-        d="M25.9999 30.32C26.8643 30.3229 27.71 30.0691 28.4301 29.591C29.1501 29.1128 29.7121 28.4317 30.0448 27.634C30.3775 26.8362 30.4661 25.9577 30.2992 25.1096C30.1323 24.2615 29.7175 23.482 29.1073 22.8699C28.4971 22.2577 27.7189 21.8403 26.8714 21.6707C26.0239 21.501 25.145 21.5866 24.3462 21.9168C23.5474 22.2469 22.8645 22.8066 22.384 23.5251C21.9035 24.2436 21.647 25.0885 21.647 25.9529C21.6458 26.5255 21.7574 27.0928 21.9756 27.6222C22.1938 28.1517 22.5141 28.633 22.9184 29.0386C23.3227 29.4442 23.8029 29.7661 24.3316 29.986C24.8604 30.2059 25.4273 30.3194 25.9999 30.32Z"
-        fill="currentColor"
-      />
-    </Icon>
-  );
-}
-
-export default ReactIcon;
