--- components/Icons/SQLIcon.tsx
@@ -1,35 +1,0 @@
-import React from "react";
-import { Icon, IconProps } from "@chakra-ui/react";
-
-function SQLIcon(props: IconProps): JSX.Element {
-  return (
-    <Icon viewBox="0 0 48 49" fill="none" {...props}>
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M34.3905 9.35417C31.8057 10.1738 28.1763 10.6836 24.1571 10.6836C19.7941 10.6836 15.8903 10.0828 13.2837 9.13717C14.2029 8.22035 18.5445 7.52612 23.7617 7.52612C29.3832 7.52612 33.9882 8.33215 34.3905 9.35417Z"
-        fill="currentColor"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M34.3906 9.35417C36.8525 8.57344 38.3665 7.51169 38.3665 6.3419C38.3665 3.94393 32.0045 2 24.1566 2C16.3087 2 9.94678 3.94393 9.94678 6.3419C9.94678 7.40679 11.2014 8.38215 13.2837 9.13748C14.2026 8.22058 18.5443 7.52628 23.7618 7.52628C29.3831 7.52628 33.9879 8.33222 34.3906 9.35417Z"
-        fill="currentColor"
-      />
-      <path
-        d="M23.565 46.9997V13.78C11.8005 13.78 6.95314 9.61123 6 7.52686V41.5282C6 42.3098 7.63395 44.2639 13.3528 45.8272C17.9278 47.0778 22.1834 46.9997 23.565 46.9997Z"
-        fill="currentColor"
-      />
-      <path
-        d="M23.5649 46.9992V13.7795C35.3294 13.7795 40.1767 9.61074 41.1299 7.52637V41.5277C41.1299 42.3093 39.4959 44.2635 33.7771 45.8267C29.2021 47.0774 24.9464 46.9992 23.5649 46.9992Z"
-        fill="currentColor"
-      />
-      <path
-        d="M11.6699 30.624V28.4365C12.0664 28.7692 12.4971 29.0199 12.9619 29.1885C13.4268 29.3525 13.8962 29.4346 14.3701 29.4346C14.6481 29.4346 14.8896 29.4095 15.0947 29.3594C15.3044 29.3092 15.4775 29.2409 15.6143 29.1543C15.7555 29.0632 15.8604 28.9583 15.9287 28.8398C15.9971 28.7168 16.0312 28.5846 16.0312 28.4434C16.0312 28.252 15.9766 28.0811 15.8672 27.9307C15.7578 27.7803 15.6074 27.6413 15.416 27.5137C15.2292 27.3861 15.0059 27.263 14.7461 27.1445C14.4863 27.026 14.2061 26.9053 13.9053 26.7822C13.1396 26.4632 12.5677 26.0736 12.1895 25.6133C11.8158 25.153 11.6289 24.597 11.6289 23.9453C11.6289 23.4349 11.7314 22.9974 11.9365 22.6328C12.1416 22.2637 12.4196 21.9606 12.7705 21.7236C13.126 21.4867 13.5361 21.3135 14.001 21.2041C14.4658 21.0902 14.958 21.0332 15.4775 21.0332C15.988 21.0332 16.4391 21.0651 16.8311 21.1289C17.2275 21.1882 17.5921 21.2816 17.9248 21.4092V23.4531C17.7607 23.3392 17.5807 23.2389 17.3848 23.1523C17.1934 23.0658 16.9951 22.9951 16.79 22.9404C16.585 22.8812 16.3799 22.8379 16.1748 22.8105C15.9743 22.7832 15.7829 22.7695 15.6006 22.7695C15.3499 22.7695 15.1221 22.7946 14.917 22.8447C14.7119 22.8903 14.5387 22.9564 14.3975 23.043C14.2562 23.1296 14.1468 23.2344 14.0693 23.3574C13.9919 23.4759 13.9531 23.6104 13.9531 23.7607C13.9531 23.9248 13.9964 24.0729 14.083 24.2051C14.1696 24.3327 14.2926 24.4557 14.4521 24.5742C14.6117 24.6882 14.8053 24.8021 15.0332 24.916C15.2611 25.0254 15.5186 25.1393 15.8057 25.2578C16.1976 25.4219 16.5485 25.5973 16.8584 25.7842C17.1729 25.9665 17.4417 26.1738 17.665 26.4062C17.8883 26.6387 18.0592 26.9053 18.1777 27.2061C18.2962 27.5023 18.3555 27.8486 18.3555 28.2451C18.3555 28.792 18.2507 29.2523 18.041 29.626C17.8359 29.9951 17.5557 30.2959 17.2002 30.5283C16.8447 30.7562 16.43 30.9202 15.9561 31.0205C15.4867 31.1208 14.9899 31.1709 14.4658 31.1709C13.9281 31.1709 13.4154 31.1253 12.9277 31.0342C12.4447 30.943 12.0254 30.8063 11.6699 30.624ZM24.1113 31.1709C22.7077 31.1709 21.5638 30.7152 20.6797 29.8037C19.7956 28.8831 19.3535 27.6914 19.3535 26.2285C19.3535 24.6836 19.8024 23.4303 20.7002 22.4688C21.5934 21.5117 22.7829 21.0332 24.2686 21.0332C25.6676 21.0332 26.7979 21.4912 27.6592 22.4072C28.5251 23.3187 28.958 24.5264 28.958 26.0303C28.958 27.5661 28.5091 28.8079 27.6113 29.7559C27.5794 29.7878 27.5498 29.8197 27.5225 29.8516C27.4951 29.8789 27.4655 29.9062 27.4336 29.9336L29.915 32.3193H26.8252L25.5264 31C25.098 31.1139 24.6263 31.1709 24.1113 31.1709ZM24.207 22.9336C23.4368 22.9336 22.8216 23.2253 22.3613 23.8086C21.9056 24.3874 21.6777 25.1553 21.6777 26.1123C21.6777 27.083 21.9056 27.8509 22.3613 28.416C22.8171 28.9811 23.4141 29.2637 24.1523 29.2637C24.9134 29.2637 25.5173 28.9902 25.9639 28.4434C26.4105 27.8874 26.6338 27.124 26.6338 26.1533C26.6338 25.1416 26.4173 24.3509 25.9844 23.7812C25.556 23.2161 24.9635 22.9336 24.207 22.9336ZM36.4297 31H30.5918V21.1973H32.7998V29.209H36.4297V31Z"
-        fill="white"
-      />
-    </Icon>
-  );
-}
-
-export default SQLIcon;
