--- components/Icons/TypeScriptIcon.tsx
@@ -1,15 +1,0 @@
-import React from "react";
-import { Icon, IconProps } from "@chakra-ui/react";
-
-function TypeScriptIcon(props: IconProps): JSX.Element {
-  return (
-    <Icon viewBox="0 0 45 45" {...props}>
-      <path
-        d="M0.260742 0.260986V44.561H44.5607V0.260986H0.260742ZM26.511 23.5209H21.2983V39.6388H17.0972V23.5209H11.9929V19.9499H26.511V23.5209ZM27.328 38.7183V34.4089C27.328 34.4089 29.6809 36.1834 32.5062 36.1834C35.3316 36.1834 35.2208 34.3375 35.2208 34.084C35.2208 31.4039 27.2198 31.4039 27.2198 25.4677C27.2198 17.3928 38.878 20.5799 38.878 20.5799L38.7328 24.4168C38.7328 24.4168 36.7787 23.1124 34.5686 23.1124C32.361 23.1124 31.5636 24.1633 31.5636 25.2856C31.5636 28.1823 39.6385 27.8919 39.6385 33.7223C39.6385 42.7004 27.328 38.7183 27.328 38.7183Z"
-        fill="currentColor"
-      />
-    </Icon>
-  );
-}
-
-export default TypeScriptIcon;
