--- components/Icons/WebflowIcon.tsx
@@ -1,15 +1,0 @@
-import React from "react";
-import { Icon, IconProps } from "@chakra-ui/react";
-
-function WebflowIcon(props: IconProps): JSX.Element {
-  return (
-    <Icon viewBox="0 0 45 45" fill="none" {...props}>
-      <path
-        d="M22.5 0C10.0732 0 0 10.0732 0 22.5C0 34.9268 10.0732 45 22.5 45C34.9268 45 45 34.9268 45 22.5C45 10.0732 34.9268 0 22.5 0ZM30.7233 26.7701C29.7059 29.1874 27.8198 31.4012 24.3293 31.4012C24.3293 31.4012 22.8492 21.0874 22.8267 20.8507C22.7358 21.0991 20.6462 26.5216 20.6462 26.5216C19.663 29.0289 17.7877 31.3895 14.2973 31.4012L11.564 14.7277C13.8346 14.7277 16.6363 16.5913 16.8398 19.664C16.8398 19.664 17.19 25.7752 17.2017 26.2614C17.3935 25.742 19.7658 19.5965 19.7658 19.5965C20.7039 17.0775 22.4775 14.717 25.8095 14.717C25.8095 14.717 27.2328 25.7195 27.2896 26.2389C27.4363 25.7195 29.3909 19.5965 29.3909 19.5965C30.328 17.0775 32.1466 14.717 35.6028 14.717L30.7233 26.7701Z"
-        fill="currentColor"
-      />
-    </Icon>
-  );
-}
-
-export default WebflowIcon;
