--- components/Icons/WebpackIcon.tsx
@@ -1,23 +1,0 @@
-import React from "react";
-import { Icon, IconProps } from "@chakra-ui/react";
-
-function WebpackIcon(props: IconProps): JSX.Element {
-  return (
-    <Icon viewBox="0 0 51 51" fill="none" {...props}>
-      <path
-        d="M25.5001 0.00854492L48.0251 12.75V38.2415L25.5001 50.9831L2.9751 38.2415V12.75L25.5001 0.00854492Z"
-        fill="white"
-      />
-      <path
-        d="M44.0044 37.3574L26.2479 47.4129V39.5929L37.3149 33.4984L44.0044 37.3574ZM45.2199 36.2609V15.2489L38.7259 18.9974V32.5124L45.2199 36.2609ZM6.92741 37.3574L24.6839 47.4044V39.5844L13.6169 33.4984L6.92741 37.3574ZM5.71191 36.2609V15.2489L12.2059 18.9974V32.5124L5.71191 36.2609ZM6.46841 13.8889L24.6839 3.58691V11.1519L13.0134 17.5694L12.9199 17.6204L6.46841 13.8889ZM44.4549 13.8889L26.2479 3.58691V11.1519L37.9099 17.5779L38.0034 17.6289L44.4549 13.8889Z"
-        fill="#585858"
-      />
-      <path
-        d="M24.684 37.8081L13.77 31.7987V19.9072L24.684 26.2057V37.8081ZM26.248 37.8081L37.162 31.8072V19.9072L26.248 26.2057V37.8081ZM14.501 18.5302L25.466 12.5037L36.4225 18.5302L25.466 24.8542L14.501 18.5302Z"
-        fill="#171923"
-      />
-    </Icon>
-  );
-}
-
-export default WebpackIcon;
