--- components/Icons/WordpressIcon.tsx
@@ -1,27 +1,0 @@
-import React from "react";
-import { Icon, IconProps } from "@chakra-ui/react";
-
-function WordpressIcon(props: IconProps): JSX.Element {
-  return (
-    <Icon viewBox="0 0 51 51" fill="none" {...props}>
-      <path
-        d="M25.7989 27.0237L22.5317 36.5563H22.5218L19.9917 43.8377C20.171 43.8875 20.3403 43.9274 20.5196 43.9772H20.5495C22.158 44.4061 23.8155 44.6238 25.4802 44.6246C26.3104 44.6308 27.1398 44.5709 27.9604 44.4453C29.0587 44.3077 30.1422 44.071 31.1977 43.7381C31.4567 43.6584 31.7157 43.5688 31.9747 43.4791C31.6958 42.8815 31.0981 41.5567 31.0683 41.4969L25.7989 27.0237Z"
-        fill="currentColor"
-      />
-      <path
-        d="M8.04844 17.978C7.05234 20.2292 6.375 22.9983 6.375 25.4985C6.375 26.1261 6.40488 26.7536 6.46465 27.3712C7.15195 34.4634 11.734 40.42 18.0293 43.0995C18.2883 43.2091 18.5572 43.3186 18.8262 43.4183L9.5625 17.988C8.76562 17.9581 8.61621 18.0079 8.04844 17.978Z"
-        fill="currentColor"
-      />
-      <path
-        d="M42.8519 17.47C42.4234 16.5474 41.9203 15.6613 41.3478 14.8204C41.1885 14.5814 41.0092 14.3423 40.8398 14.1032C38.6815 11.2063 35.7505 8.97622 32.383 7.66848C30.1866 6.80934 27.8485 6.37009 25.49 6.37356C22.633 6.37362 19.8125 7.01543 17.2367 8.25158C14.661 9.48774 12.3958 11.2867 10.6084 13.5155C9.96091 14.3182 9.37803 15.1708 8.86523 16.0655C10.2797 16.0755 12.0328 16.0755 12.232 16.0755C14.035 16.0755 16.8141 15.8564 16.8141 15.8564C17.7504 15.7966 17.85 17.1612 16.9236 17.2708C16.9236 17.2708 15.9873 17.3804 14.9514 17.4302L21.2168 36.0571L24.982 24.7714L22.3125 17.4402C21.3762 17.3903 20.5096 17.2808 20.5096 17.2808C19.5732 17.231 19.6928 15.8066 20.6092 15.8663C20.6092 15.8663 23.448 16.0855 25.1414 16.0855C26.9443 16.0855 29.7234 15.8663 29.7234 15.8663C30.6498 15.8066 30.7693 17.1712 29.833 17.2808C29.833 17.2808 28.9066 17.3903 27.8707 17.4402L34.0764 35.9276L35.7996 30.1901C36.6662 27.9589 37.1045 26.1161 37.1045 24.6419C37.1045 22.5202 36.3375 21.046 35.6801 19.9005C34.8135 18.4761 33.9967 17.2808 33.9967 15.8763C33.9967 14.2925 35.192 12.8183 36.8853 12.8183H37.1045C39.7143 12.7486 40.5709 15.3384 40.6805 17.1015V17.1612C40.7203 17.8784 40.6904 18.4064 40.6904 19.0339C40.6904 20.7671 40.3617 22.7294 39.3855 25.1898L35.5008 36.4257L33.2795 42.97C33.4588 42.8903 33.6281 42.8106 33.8074 42.721C39.4553 39.9917 43.5691 34.553 44.4357 28.1083C44.5642 27.2511 44.6275 26.3853 44.625 25.5185C44.6288 22.7384 44.0236 19.9912 42.8519 17.47Z"
-        fill="currentColor"
-      />
-      <path
-        d="M25.5 4.77954C30.2801 4.78201 34.9124 6.43676 38.6121 9.46348C42.3118 12.4902 44.8516 16.7028 45.8011 21.3876C46.7506 26.0725 46.0514 30.9415 43.8221 35.1699C41.5927 39.3983 37.9703 42.726 33.5684 44.5894C28.5051 46.7293 22.7991 46.7701 17.7057 44.7029C12.6123 42.6358 8.54873 38.6299 6.40887 33.5666C4.269 28.5034 4.22817 22.7974 6.29534 17.704C8.36251 12.6106 12.3684 8.54702 17.4316 6.40716C19.9841 5.32709 22.7284 4.77349 25.5 4.77954ZM25.5 3.18579C13.1773 3.18579 3.1875 13.1756 3.1875 25.4983C3.1875 37.821 13.1773 47.8108 25.5 47.8108C37.8227 47.8108 47.8125 37.821 47.8125 25.4983C47.8125 13.1756 37.8227 3.18579 25.5 3.18579Z"
-        fill="currentColor"
-      />
-    </Icon>
-  );
-}
-
-export default WordpressIcon;
