+++ components/Icons/index.ts
@@ -1,0 +1,3 @@
+export { default as GithubIcon } from "./GithubIcon";
+export { default as InstagramIcon } from "./InstagramIcon";
+export { default as TwitterIcon } from "./TwitterIcon";
