+++ components/Menu.tsx
@@ -2,7 +2,8 @@ import { ExternalLinkIcon } from "@chakra-ui/icons";
 import { Button, Flex, useBreakpointValue } from "@chakra-ui/react";
 import Link from "next/link";
 import React from "react";
-function Menu(): JSX.Element {
+
+export default function Menu(): JSX.Element {
   const btnSize = useBreakpointValue({ base: "sm", sm: "md" });
   return (
     <Flex justifyContent="space-between">
@@ -14,9 +15,9 @@ function Menu(): JSX.Element {
         </Link>
       </div>
       <div>
-        <Link href="/tech">
+        <Link href="/recommendations">
           <Button colorScheme="blue" size={btnSize} variant="ghostCyan">
-            Tech
+            Recs
           </Button>
         </Link>
       </div>
@@ -35,4 +36,3 @@ function Menu(): JSX.Element {
     </Flex>
   );
 }
-export default Menu;
