+++ components/SolidDivider.tsx
@@ -1,6 +1,7 @@
 import { Divider, DividerProps, useColorModeValue } from "@chakra-ui/react";
 import React from "react";
-function SolidDivider({ ...props }: DividerProps): JSX.Element {
+
+export default function SolidDivider({ ...props }: DividerProps): JSX.Element {
   const borderColor = useColorModeValue("gray.900", "gray.100");
   return (
     <Divider
@@ -11,4 +12,3 @@ function SolidDivider({ ...props }: DividerProps): JSX.Element {
     />
   );
 }
-export default SolidDivider;
