+++ components/index.ts
@@ -1,1 +1,4 @@
 export { default as Music } from "./Music";
+export { default as Header } from "./Header";
+export { default as Layout } from "./Layout";
+export { debounce } from "./helpers";
