+++ config/index.ts
@@ -1,0 +1,2 @@
+export { default as ChakraWrapper } from "./chakra";
+export { default as MDXWrapper } from "./mdx";
