--- pages/blog/howcome.mdx
@@ -1,11 +1,0 @@
-### 👍🏽 Recommendations
-
-Here's a collection of recommendations that I use on a regular basis.
-
-This list is inspired by @AlexCWatt's list
-
-#### Talks
-
-- Something about software
-
-#### Software
