+++ pages/index.mdx
@@ -1,4 +1,4 @@
-### Hi 🫡
+### Hi
 
 I'm a developer based out of beautiful [Sacramento](https://www.google.com/maps/place/Sacramento,+CA/@38.5616505,-121.582995,11z/data=!3m1!4b1!4m5!3m4!1s0x809ac672b28397f9:0x921f6aaa74197fdb!8m2!3d38.5815719!4d-121.4943996).
