+++ pages/recommendations.mdx
@@ -1,3 +1,11 @@
-### Recommendations
+### 👍🏽 Recommendations
 
-- Something cool about
+Here's a collection of recommendations that I use on a regular basis.
+
+This list is inspired by @AlexCWatt's list
+
+#### Talks
+
+- Something about software
+
+#### Software
