--- pages/tech.mdx
@@ -1,7 +1,0 @@
-## Tech I use
-
-One of the biggest inspirations is keeping a list of things that I'm consistently using over a regular basis
-
-### Talks
-
-### VSCode Editor
